[adapters] Implement suspend and resume without fault tolerance.
A big part of this commit is rearranging code in the controller, moving
code out of the fault tolerance-specific paths in `FtState` into code
that is used when fault-tolerance is not enabled.

The new `CanSuspend` type added to global metrics is there so that UX
code can decide whether to show the potential for a suspend operation
(for Yes or NotNow) or not (for No).  For example, Yes might be a usable
suspend button, NotNow a grayed-out suspend button, and No no suspend
button at all.

Also adds a test.

Signed-off-by: Ben Pfaff <blp@feldera.com>