It compiles.
The new strategy is:

Input connector will read data and call stage() on each InputBuffer as it
goes. After it's staged enough, it calls Parser::gather_staged(), which
produces a StagedBuffers.  When it's time to queue, it creates a new
StagedBuffers (if there isn't already one) and then calls flush_staged()
on the StagedBuffers (either the new one or one that it already had).